Made RandRedirector into internal class inside of PatchingUtilities
